Fix Access of Minecraft Class and add Launch Methods
Minecraft classes were originally all in the same package, but with the
new mappings we need to change some access modifiers. When the jars are
merged together, we change all protected/package-private entities to
public.

One little bug with the TSRG was fixed. It wasn't adding the last class
of a mapping to the list of classes.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>