Remap Obfuscated Server Methods & Fields to Client Obfuscation
Instead of an ASM Remapper, we need to use SS because it does a great
job of respecting inheritence.

There's a lot of debug code in here that I will remove soon. I'm
committing now because I need to do some major reforms in the next
commit since PACKET CONSTRUCTORS DON'T EXIST CLIENTSIDE FOR SOMEREASONLIKEWHATTHE
HELL IS HAPPENINGHOWDO THEY EVENDEPLOY THIS CRAP?@??@#?$?@

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>